Dungeon Framework
Dungeon should be *mostly* complete. I need to rework the toString (it's just there to test it out), and the items need to be built. It'll need a bit more work once we get the game interface going and need dialogue/removing items from the room when the player picks them up.