Started MVC
Started the MVC, as well as the core gameplay loop. Parts of it are commented out so they don't pull errors for the time being. The StartView class is the VIEW class for ONLY the game start. It will not be used after the game is initialized. All output will be sent to GAMEVIEW afterwards. I also put more time into the dungeon class; It's distributes an exit and an entrance now.